+++ CHANGELOG.md
@@ -28,6 +28,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 ### Deprecated
 
 ### Removed
+- Drop the automatic gender-specific salutations (#2288)
 - Drop the changelog archive (#2284)
 - Drop the legacy unregistration form (#2278)
 - Drop the `showOwnerDataInSingleView` setting (#2267)